[Extensions WebIDL] Remove 'type': 'promise' in 'returns_async'
Although this would be a nice addition and would be useful for merging
the 'returns' and 'returns_async' fields, for now while converting
schemas to WebIDL it makes sense to not include this yet so it is easier
to verify that there is no difference between the old IDL processor and
the new WebIDL processor.

crbug.com/428187556 has been filed to come back to this after the move
to WebIDL is complete.

Bug: 340297705
Change-Id: I86f1f7fb1639fbf6d0933dd29c7dcdbb551792aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6686226
Commit-Queue: Tim <tjudkins@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480693}