Reclassify initial MHTML fragment navigation as cross-document.
Reclassified the initial navigation from the MHTML empty document to a
fragment (e.g., about:blank#foo) as a cross-document navigation. This
change aligns with the fact that the initial fragment navigation in
MHTML commits a new document in a new opaque origin.

Subsequent fragment navigations (e.g., #foo to #bar) are still treated
as same-document, as expected.

The change is guarded by the feature flag
TreatMhtmlInitialDocumentLoadsAsCrossDocument to allow for safe rollout.

Bug: 423663315
Change-Id: I0bc71c5972e6264c9abbde62dff44f38f937181a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6653528
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Commit-Queue: Diana Qu <xiaqu@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1480799}