[Extensions WebIDL] Support extended attributes above Operations
Previously if an extended attribute was included on the line above an
Operation definition it would mess up the comment parsing. This was due
to an exception case we already had to do for Operations to deal with
them not having correct line number annotations and instead look at the
line numbers on their arguments.

This CL fixes that by instead doing the extended attribute check first,
and then falling back to the arguments check if there are no extended
attributes.

Also adds related test and deletes the related TODOs.

Bug: 340297705
Change-Id: I903c129d05d6955f1a3927792caa90303df56871
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6692001
Reviewed-by: Justin Lulejian <jlulejian@chromium.org>
Commit-Queue: Tim <tjudkins@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480805}