pin-infobar: add controller stub
This change adds a stub (i.e., non-implemented) version of
`PinInfoBarController`. This is a preliminary step toward adding an
infobar that will prompt the user to pin Chrome to the taskbar.

See crrev.com/c/6642547 for the full implementation, which will follow
this CL as a dependent CL.

Bug: 420960161
Change-Id: If329b23b8bf83f976810f8e125566fd088c16131
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6669299
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Reviewed-by: Darryl James <dljames@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480824}