extensions: Compile ExtensionUninstallDialogView on Desktop Android
Now that ShowDialog() is available, ExtensionUninstallDialogView mostly
works on Desktop Android. This patch moves it out of //c/b/ui/views/...
to make it clear that it is not views-specific, and rename it to
ExtensionUninstallDialogImpl.

ModalDialogWrapper still lacks checkbox support, so we #ifdef-out it
temporarily. The checkbox supported is tracked at crbug.com/404069731.

Bug: 424011073
Change-Id: I43860f15b20f72e6a997c0819e1ecfe9d52a263b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6656243
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480851}