[overscroll] Enable ElasticOverscroll flag on non-mac platforms.
Enable elastic overscroll to be tested and debugged on all platforms.
This will allow for faster development of supporting elastic overscroll
on all scrollable sublayers.

Returning false from IsElasticOverscrollEnabled is not sufficient to
disable elastic overscroll which is why this flag is not included on
mac.

Bug: 41102897
Change-Id: I92660d9583fa79cebc42cca082f00ba4166518f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6690337
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480880}