Split ScreenCapture.ImageHandler into a separate class.
This CL moves ScreenCapture.ImageHandler to a separate file. It is
currently a non-static inner class which can be hard to reason about.

To move it out, we define a Delegate to handle the coupling between
ScreenCapture and ImageHandler. We also change ImageHandler to use
internal locking. This changes the locking regions slightly - for
example, previously in onRgbaFrameAvailable if the native side was
already destroyed we wouldn't try to acquire the image. Now the
destruction locking logic and skipping of JNI calls is delayed until it
gets into ScreenCapture code. This is okay, though.

Bug: 352187279
Test: CQ
Change-Id: Icc5fb29a3e9939e95e05512082ae95b7fb44927f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6651995
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Eliot Courtney <edcourtney@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480893}