PrefetchReusable: Remove feature flag (Part 1/4)
This CL removes use of the feature flag `kPrefetchReusable` in tests as
much as possible. The feature parameter `kPrefetchReusableBodySizeLimit`
in a test will be handled in a follow-up CL [1].

No behavioral change.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/6650560

About a failure of fuchsia-x64-cast-receiver-rel:

- https://luci-milo.appspot.com/ui/inv/task-chromium-swarm.appspot.com-71d773f3cd414b11/test-results
- https://chromium-swarm.appspot.com/task?id=71d773f3cd414b10

Notice that 1. 0 PASSED, 0 SKIPPED, 0 FAILED. 2. Exit ocde is 123 and
the log contains

```
[I2025-06-29T19:24:09.567444-07:00 235334 0 cmd_stream.go:492] rdb-stream: exiting with 123
```

3. This CL only reduces tests.

So, I believe that it's an infra failure and not relevant to the CL.

Bug: 397882995
Validate-Test-Flakiness: skip
Change-Id: If3d7e223ef73457809ef9527474530f4d8762f57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6650559
Reviewed-by: Hiroshige Hayashizaki <hiroshige@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Taiyo Mizuhashi <taiyo@chromium.org>
Commit-Queue: Ken Okada <kenoss@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1480950}