[autofill_ml] Create ML prediction log router
It is provided by a browser context keyed factory.

The page handler registers itself as an observer when a tab of
autofill-ml-internals is opened and unregisters itself when the tab is
closed. The router can then know if there are any tabs open, and we can
avoid constructing and passing large log objects if we don't need to.

Bug: 428686605

Change-Id: Ia609e54a86ac8216e0652cf7d57cc379ff502186
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6645682
Reviewed-by: Ryan Sultanem <rsult@google.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Teo Lamort de Gail <lamort@google.com>
Cr-Commit-Position: refs/heads/main@{#1480982}