[iOS] Prevent crash from invalid tab index during swipe.
This change resolves an issue where GetWebStateAt was called with an
invalid index during tab-swiping gestures.
This typically occurred when the WebStateList was updated at the same
time the swipe is cancelled.
Repro case:
- Open several NTP.
- Activate the last one.
- Start the side swipe.
- Without releasing it, with your other hand go to the home screen.

The fix consists on removing the cancelTabSwitch method as it is
proven to be unnecessary.

Bug: 428174753
Change-Id: I97e39935c909b3247a28916ec5873aaf13bc0221
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6682130
Auto-Submit: Ameur Hosni <ameurhosni@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Ameur Hosni <ameurhosni@google.com>
Cr-Commit-Position: refs/heads/main@{#1481034}