[ios] Check NewTabPageTabHelper existence before use
When CreateTabHelperOnlyForRealizedWebStates() is defined, TabHelpers
will not be created for unrealized WebStates.

Check that NewTabPageTabHelper exists when iterating over WebStates
in a WebStateList to determine whether the NTP coordinator should be
started or stopped.

Unrealized WebState should not be considered to make this decision,
so treat a missing NewTabPageTabHelper as if IsActive() was false.

Bug: 409299519
Change-Id: I3fff782c9e11a4225abcb3142132be2d8bef6f2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6696644
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1481042}