[ios] Activate the WebState before starting the NTP
When CreateTabHelperOnlyForRealizedWebStates() is defined, TabHelpers
will not be created for unrealized WebStates.

When selecting a WebState to become the new NTP and starting it, mark
it as active (which will force it to become realized) before accessing
the NewTabPageTabHelper. The activation will ensure that the object is
created and can be accessed.

Bug: 409299519
Change-Id: I011563d5e91ebd839a74cb08bec67cf04eb4c9f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6696008
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1481043}