Add logs for normal usage of Catap.
The Catap file have logs that shows errors, but not for usage. So it
would be impossible to to verify that the code was running, unless an
error was triggered.

This Add logs for normal usage, so we can see in the logs that the code
was running.

Change-Id: I98741563b5f7c562f03fcdc6daa4216eb334ae6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6696344
Reviewed-by: Fredrik Hernqvist <fhernqvist@google.com>
Commit-Queue: Olov Brändström <brandstrom@chromium.org>
Reviewed-by: Henrik Andreasson <henrika@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1481075}