[Save and Fill] Offer local Save and Fill
Branch out the save and fill manager logic to offer the local Save and
Fill dialog. Added a callback that includes the user provided data
(which will be extracted from the dialog) as well as the user decision.

Local save implementation by the end of the CL chain:
http://screencast/cast/NjY5NzQ1MDc0MzAwNTE4NHwyNWQzZDBiNy1lMg

Bug: 378164516
Change-Id: I942d6ef467dc348187011ce8e954260b68de9660
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6658426
Commit-Queue: Verina Armanyous <averina@google.com>
Reviewed-by: Siyu An <siyua@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1481089}