[Trusted Types] Implement Element.inner/outerHTML without StringContext.
The [StringContext=] extendend IDL attribute is a Chrome-specific
mechanism to implement Trusted Types. TT is currently being integrated
into the HTML spec, and the spec editors decided against the
StringContext IDL attribute, and instead specify Trusted Types checks
"manually" in the spec. In some instances -- but not here -- the
difference between the IDL-based and the "manual" TT checks is observable, due to the when the TT check is being made.

This CL re-implements the Trusted Types check of the following properties, without StringContext:
- Element, innerHTML + outerHTML
- Shadow root, innerHTML

These methods are very popular in testing code, which unfortunately
makes this CL quite long.

This CL should not introduce any behavioural changes. If the updated
HTML specs introduces changes relative to Chrome's current behaviour,
we will implement those semantic changes in a separate CL, behind a
flag. Here, we only change the mechanism.

Bug: 330516530
Change-Id: I5e3de3a7f8dec469a27314056958a7b982e8733d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6624252
Commit-Queue: Daniel Vogelheim <vogelheim@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1481115}