Set initial focus inside the Plus Address Creation dialog
When the Plus Address Creation dialog(popup) opens, focus is not
transferred into an element inside the dialog. As a result, users using
keyboard controls cannot access or close the dialog.

This change moves the initial focus to the 'cancel' button in the dialog
while a plus address is being reserved. Once it's reserved, it sets
focus to the 'use address' button. The majority of the time the reserve
will not take long enough for the focus on the `cancel` button to be
show in the ui or read by the screen reader. The user can then use
keyboard controls to traverse the dialog or close it.

Demo: http://screencast/cast/NDY2MjA2MzcyODE2NDg2NHw2YWE4M2FlOS00ZQ
Bug: crbug.com/408234477
Change-Id: I4b5e64c2cabc3953e35ffdb8396d94a795c64db2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6680952
Reviewed-by: Jan Keitel <jkeitel@google.com>
Reviewed-by: Matt Reichhoff <mreichhoff@chromium.org>
Commit-Queue: Fikre Mengistu <fikremengistu@chromium.org>
Reviewed-by: Kirubel Aklilu <kaklilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1481130}