[remoting] Remove unnecessary call to IgnoreXServerGrabs().
In the remoting host, IgnoreXServerGrabs() is called in 3 places:

1. BasicDesktopEnvironment ctor
2. LegacyInteractionStrategy, before each screen-capturer is created
3. InputInjectorX11

In places 1 and 2, the call operates on the libX11 Display which is
part of the webrtc::DesktopCaptureOptions. Those Displays are created
and owned by the WebRTC desktop-capture module. In place 3, the call
operates on an x11::Connection object, which is the Chromium
implementation of the X11 protocol, separate from libX11.

In place 1, the libX11 Display actually gets created during startup in
the HostProcess ctor which calls through to CreateDefault() in WebRTC.
But each screen-capturer runs on a separate libX11 Display - see
http://crrev.com/c/3898352. So this Display is not used by anything.
All other X11 calls either use the Chromium x11::Connection, or some
separate X11 connection inside GTK/GDK.

This CL removes the call from place 1. Of the 3 places, only place 1
was guarded by a watchdog. This watchdog is deprecated in favor of
HangWatcher, and is probably not needed since we initialized libX11
for multi-threaded use in http://crrev.com/c/2780005. So, instead of
moving the watchdog to place 2, we remove it completely.

Removing this code also has the benefit of not calling it if Wayland
is enabled.

Bug: 40149664
Change-Id: I9e2f86abf2fe37fc4ee78baa8cc4238d1fca6f16
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6692991
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1481170}