[CLAMS] Skip on device inquiry if inner text is too short
When the inner text is too short, it can cause hallucination in the on
device model output. Therefore, we should skip the inquiry so that the
resources are saved. We have settled on 5 as the threshold based on
manual analysis with internal data.

Server side update: cl/777755661

Bug: 393197678
Change-Id: I5218faaf9facb106ec54f8fd1eadcd57174876ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6692351
Commit-Queue: Seung Jae Lim <andysjlim@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1481188}