Migrating SafeBrowsingUI structs to util file
The SafeBrowsingUI file has hundreds of lines of code contained in
private functions. As outlined in the SafeBrowsingUI refactor document,
a new util file is to be created and filled with these private functions
as well as other structs that are shared across different classes used
in SafeBrowsingUI.

Since amount of code that will be migrated to the new util file is
large, the migration will take place in steps. The first step, which
this CL covers, is to migrate the shared struct objects.

This is a multi-CL change:
  => https://crrev.com/c/6621108
  -  https://crrev.com/c/6638271
  -  https://crrev.com/c/6637992

Bug: 427719727
Change-Id: If1a6d691bb49e3db06c7fea2c14dc8ccf1318533
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6621108
Reviewed-by: thefrog <thefrog@chromium.org>
Commit-Queue: Daniel White <danieltwhite@google.com>
Cr-Commit-Position: refs/heads/main@{#1481189}