[iOS] Stop showing Home Customization 'new' badge after menu is tapped
This change updates the Feature Engagement Tracker (FET) configuration
for the kIPHiOSHomepageCustomizationNewBadge feature. The configuration
now uses the existing events::kHomeCustomizationMenuUsed event to
control the visibility of the 'new' badge. This ensures that the badge
is no longer displayed after the user has tapped the customization menu
for the first time.

Bug: 428925379
Change-Id: I6fd18895eb574d3bcc7e157fef0f6fbaec54dd9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6696904
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1481200}