[iOS] Set NTPViewController's mutator earlier in coordinator lifecycle
This change refactors the NewTabPageCoordinator to set the mutator on
the NTPViewController earlier in the coordinator lifecycle. Previously,
the mutator was set during the -configureNTPViewController method. It is
now set immediately after the NTPViewController is initialized. This
ensures the mutator is available for the entire lifecycle of the view
controller, preventing potential nil access or timing issues.

Bug: 428928357
Change-Id: I8c8e6294e7034164851dc6d2d4b9420adf8d3b3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6697093
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Reviewed-by: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1481202}