Fix NPE in AutoCompleteMediator
This was a longstanding bug, but the signature changed due to the recent
null annotations and this got assigned to me. If a tab doesn't have a
window it seems sensible to just early out since it won't be found in a
tab model if it isn't attached.

Fixed: 428931424
Change-Id: I898cf3bed2a6c072f69f435799db301dd97abe09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6697073
Auto-Submit: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Sky Malice <skym@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1481212}