[PWA] FetchManifestAndInstallCommand now uses the install info creation
This CL migrates the FetchManifestAndInstallCommand to now use the
ManifestToWebAppInstallInfoJob. To do so, the command has been modified with the following changes:

1. Once the manifest is obtained, all necessary configurations for icon fetching are stored in a single swoop instead of being stored all over the place.
2. Setting whether the app is DIY or not is now streamlined,
and happens only in a single place, once the manifest has been parsed
and converted to a WebAppInstallInfo. This makes reasoning about the
whole flow simpler.
3. A new function to parse an existing web app
install info obtained from the web page metadata and merge with one
obtained from the manifest is added to support icon generation once
the title has been added to the info.

Tested manually by testing with different sites (pokedex.org and
Wikipedia), as well as by running the corresponding unit and browser
tests for the command.

Bug: 427565907
Change-Id: I394526c7c90a77a2dfb8b9da4846a3a925c7e624
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6691117
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1481319}