Bugfix: Check the result of FindFromLiveRegistrationsForId() correctly.
Depending on its internal state,
`ServiceWorkerRegistry::FindFromLiveRegistrationsForId()` can return one
of three value types:

1. `std::optional` containing a valid pointer.
2. `std::optional` containing a `nullptr`.
3. `std::nullopt`.

The previous implementation in `FindRegistrationForClientUrl()` used the
condition `if (registration)` to check the result. This condition
evaluates to `true` as long as the `std::optional` is not
`std::nullopt`, which caused the block to execute even when the optional
contained a `nullptr`.

As a result, when `FindFromLiveRegistrationsForId()` returned an
optional containing `nullptr`, it caused a bug.

This CL fixes the issue by changing the condition to `if
(registration.has_value() && registration.value())`. This ensures the
early-return path is taken only when the std::optional contains a valid
(non-null) pointer, and correctly falls through to perform the storage
lookup for `nullptr` or `std::nullopt` cases.

Bug: 429055859
Change-Id: I0da227a99b6744ea74169f22a0a8c2cb5a721572
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6698073
Commit-Queue: Minoru Chikamune <chikamune@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1481383}